requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Medical Imaging
nibabel>=3.2.0
SimpleITK>=2.1.0
dicom2nifti>=2.3.0
pydicom>=2.3.0

# nnUNet
nnunetv2>=2.0.0

# Data Processing
pandas>=1.3.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# API & Download
synapseclient>=2.7.0
requests>=2.28.0
openai>=1.0.0

# Utilities
python-dotenv>=0.19.0